docs : 문서 파일 구조 변경으로 인한 삭제
