+++ src/@components/@common/index.ts
@@ -1,0 +1,4 @@
+export { default as BackPageNav } from './BackPageNav';
+export { default as UserInputBox } from './UserInputBox';
+export { default as LearningHeader } from './LearningHeader';
+export { default as UserProfile } from './UserProfile';
